requirements.txt
# Core ML
torch==2.4.0
timm==1.0.25
einops==0.8.2

# Training (fine-tuning Clay with CDL labels)
terratorch==1.2.2
lightning==2.6.1

# Geospatial
pystac-client==0.9.0
rasterio==1.4.4
geopandas==1.1.2
shapely==2.1.2
rasterstats==0.20.0

# Data
pyarrow==23.0.1
pandas==2.3.3
numpy==2.4.2
scikit-learn==1.8.0

# CLI + logging
tqdm==4.67.3
rich==14.3.3
click==8.3.1
pyyaml==6.0.2

# Dashboard
streamlit==1.54.0
streamlit-keplergl==0.3.0
keplergl==0.3.2
plotly==6.5.2

# HTTP
requests==2.32.3
